useStateでフォームの値を管理する
コンポーネント名が小文字始めり(`signup`)の場合、lintエラー(`Rules of Hooks`)が出る

```
The error message you're seeing indicates that the useState hook is being used inside a function called signup that is not a valid React function component or a custom React Hook function.

To resolve this issue, you need to ensure that the signup function is defined as a valid React function component. React function components are typically defined with an uppercase letter at the beginning of the function name.
```