入力に応じてステートを更新する
To address the TypeScript error "Parameter 'e' implicitly has an 'any' type," explicitly specify the type of the event parameter in the handleOnSubmit function.

the handleOnSubmit function now explicitly specifies the type of the event parameter as React.FormEvent<HTMLFormElement>. This type indicates that the event is related to a form element.